TEST: moved checks for banned primitives into certain folder, out of pkgs
